// src/generated/models/medical-history-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Healix Clinic Management API
 * REST API for clinic management system with EHR support for Maharashtra small to mid-sized clinics.  ## Features - Patient registration and management - Emergency contact management - Insurance details tracking - Medical history records - Appointment scheduling and management - Vitals recording and examination - Prescription and investigation management - IPD/OPD workflow 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@healix.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface MedicalHistoryRequest {
    'knownAllergies'?: Array<string>;
    'currentMedications'?: Array<string>;
    'pastSurgeries'?: string;
    'chronicConditions'?: string;
    'familyMedicalHistory'?: string;
    'disability'?: string;
}